requirements.txt
Flask==2.3.2
Pillow==9.5.0
numpy==1.26.4
opencv-python==4.10.0.84
